+++ backend/src/main/java/cs251/group9/backend/repository/PlayedRepository.java
@@ -10,14 +10,14 @@ import java.util.*;
 public interface PlayedRepository extends JpaRepository<Played, PlayedId> {
     
     @Query(value = "SELECT * FROM played WHERE gameID = :gameID", nativeQuery = true)
-    List<Played> findByGameGameID(@Param("gameID") Integer gameId);
+    List<Played> findByGameGameID(@Param("gameID") Long gameId);
     
     @Query(value = "SELECT * FROM played WHERE userID = :userID", nativeQuery = true)
-    List<Played> findByUserID(@Param("userID") Integer userId);
+    List<Played> findByUserID(@Param("userID") Long userId);
     
     @Query(value = "SELECT COUNT(*) > 0 FROM played WHERE userID = :userID AND gameID = :gameID", nativeQuery = true)
-    boolean existsByUserIDAndGameID(@Param("userID") Integer userID, @Param("gameID") Integer gameID);
+    boolean existsByUserIDAndGameID(@Param("userID") Long userID, @Param("gameID") Long gameID);
     
     @Query(value = "SELECT * FROM played WHERE userID = :userID AND gameID = :gameID", nativeQuery = true)
-    Optional<Played> findByUserIDAndGameID(@Param("userID") Integer userID, @Param("gameID") Integer gameID);
+    Optional<Played> findByUserIDAndGameID(@Param("userID") Long userID, @Param("gameID") Long gameID);
 }
